requirements.txt
fastapi>=0.115,<1.0
uvicorn>=0.34,<1.0
pydantic>=2.10,<3.0
pydantic-settings>=2.7,<3.0
python-multipart>=0.0.18,<1.0
numpy>=2.2,<3.0
opencv-python-headless>=4.11,<5.0
pdf2image>=1.17,<2.0
Pillow>=11.1,<12.0
aiofiles>=24.1,<25.0
pytesseract>=0.3,<1.0
httpx>=0.28,<1.0
